move scripted fields from native query up to base query
Closes #2477
